logout confirm popup

+++ src/app/layout/header/header.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit, EventEmitter, Output } from '@angular/core';
 import { MatSidenav } from '@angular/material/sidenav';
 import { Router } from '@angular/router';
-import { CurrentUserService } from 'src/app/core';
+import { CurrentUserService, UtilsService } from 'src/app/core';
 
 @Component({
   selector: 'app-header',
@@ -13,7 +13,8 @@ export class HeaderComponent implements OnInit {
   @Output() toggle = new EventEmitter();
   constructor(
     private router : Router,
-    private userService : CurrentUserService
+    private userService : CurrentUserService,
+    private utilService : UtilsService
   ) { }
 
   ngOnInit(): void {
@@ -23,8 +24,16 @@ this.toggle.emit();
   }
  action(item:any){
   if(item == 'logout'){
-    this.userService.deleteUser().then(resp =>{
-      this.router.navigate(['/login']);
+    let data = {
+      header:"",
+      message:"Are you sure, you want to logout?"
+    }
+    this.utilService.openDialog(data).then(resp =>{
+      if(resp){
+        this.userService.deleteUser().then(resp =>{
+          this.router.navigate(['/login']);
+        })
+      }
     })
   }else if(item == 'profile'){
     this.router.navigate(['/profile']);
